requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
pymongo==4.5.0
marshmallow==3.20.1
python-dotenv==1.0.0
gunicorn==21.2.0